remove unused dmesg helper functions
