--- pkg/dmesg/testdata/dmesg-time-format-short-iso.log
@@ -1,11 +1,0 @@
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: docker0: port 1(vetha17b2cc) entered disabled state
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: docker0: port 1(vetha17b2cc) entered disabled state
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: device vetha17b2cc left promiscuous mode
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: docker0: port 1(vetha17b2cc) entered disabled state
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: docker0: port 1(veth9d4400a) entered blocking state
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: docker0: port 1(veth9d4400a) entered disabled state
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: device veth9d4400a entered promiscuous mode
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: IPv6: ADDRCONF(NETDEV_UP): veth9d4400a: link is not ready
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: docker0: port 1(veth9d4400a) entered blocking state
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: docker0: port 1(veth9d4400a) entered forwarding state
-2025-01-02T15:20:12+0800 jmsph-pro-a800-77-123 kernel: IPv6: ADDRCONF(NETDEV_CHANGE): veth9d4400a: link becomes ready
